REGRESSION (r169241) Crash running run-jsc-stress-tests on EFL ARM https://bugs.webkit.org/show_bug.cgi?id=133223
Reviewed by Geoffrey Garen.

Change runAndMonitorTestRunnerCommand() to convert the *cmd argument to a
string before calling IO.popen() to work with older versions of ruby.
Changed the remote path in runShellTestRunner to quote the command
string we want to run via ssh so that it will work correctly with array
to string conversion needed for popen.

* Scripts/run-jsc-stress-tests:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169265 268f45cc-cd09-0410-ab3c-d52691b4dbfc