Caret's screen position does not update during an overflow scroll https://bugs.webkit.org/show_bug.cgi?id=133256
Reviewed by Simon Fraser.

Source/WebCore:
During a scroll, we set a flag which caret position is gated on, but we never
actually invalidate the caret's position. This patch does so.

Test: editing/input/caret-position-during-scroll.html

* editing/FrameSelection.h:
* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::scrollTo):

LayoutTests:
This test places the caret, inspects its position, scrolls, and then immediately inspects its
position again. The two positions should be different.

* editing/input/caret-position-during-scroll-expected.txt: Added.
* editing/input/caret-position-during-scroll.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169399 268f45cc-cd09-0410-ab3c-d52691b4dbfc