AX: [ATK] Children of tables and lists with role='presentation' should each be exposed as if they were div elements https://bugs.webkit.org/show_bug.cgi?id=133133
Reviewed by Chris Fleizach.

Source/WebCore:
When a table or list has role='presentation', accessibility is "ignored"
for that element, but not for the required children. The default roles
for the children (CellRole and ListItemRole) do not make sense without
the parent, therefore map them to DivRole in ATK if they inherit the
presentational role.

Test: accessibility/aria-inherits-presentational.html

* accessibility/AccessibilityObject.h:
(WebCore::AccessibilityObject::inheritsPresentationalRole):
* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::inheritsPresentationalRole):
* accessibility/AccessibilityRenderObject.h:
* accessibility/atk/WebKitAccessibleWrapperAtk.cpp:
(atkRole):

LayoutTests:
Added a new test with platform-specific expectations due to differences
in accessible hierarchies.

* accessibility/aria-inherits-presentational.html: Added.
* platform/gtk/accessibility/aria-inherits-presentational-expected.txt: Added.
* platform/mac/accessibility/aria-inherits-presentational-expected.txt: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169631 268f45cc-cd09-0410-ab3c-d52691b4dbfc