AX: Crash at com.apple.WebCore: WebCore::AlternativeTextController::removeDictationAlternativesForMarker + 43 https://bugs.webkit.org/show_bug.cgi?id=134226
Reviewed by Enrica Casucci.

Source/WebCore:
Unchecked access of details pointer in the Document marker led to the crash.
It seems erroneous to allow creation of a DocumentMarker without details when one is expected, because it
carries the dictation context (not just the description).
Also, add in ASSERTs to catch scenarios where a nullptr might be passed in.

Test: platform/mac/editing/input/crash-for-empty-text-alternative.html

* dom/DocumentMarker.cpp:
(WebCore::DocumentMarker::DocumentMarker):
* editing/AlternativeTextController.cpp:
(WebCore::AlternativeTextController::removeDictationAlternativesForMarker):

LayoutTests:
* platform/mac-wk2/TestExpectations:
    Skip test on wk2 because dictation related tests are not yet supported (89401)
* platform/mac/editing/input/crash-for-empty-text-alternative-expected.txt: Added.
* platform/mac/editing/input/crash-for-empty-text-alternative.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170429 268f45cc-cd09-0410-ab3c-d52691b4dbfc