Remove an useless check from SelectorChecker https://bugs.webkit.org/show_bug.cgi?id=134868
Reviewed by Darin Adler.

* css/SelectorChecker.cpp:
(WebCore::SelectorChecker::matchRecursively):
The condition of this if() branch can never be met for the mode "QueryingRules".

The next condition in that if() is "dynamicPseudo != NOPSEUDO", which implies
a pseudo element was matched prior to the current context/simple selector.
This cannot happen with QueryingRules, since we never match pseudo elements for
SelectorQuery.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@171058 268f45cc-cd09-0410-ab3c-d52691b4dbfc