Adjustments to CueBox CSS Width calculations Part 2. https://bugs.webkit.org/show_bug.cgi?id=135820 <rdar://problem/17954473>.
Reviewed by Brent Fulgham.

Two adjustments made:
a) The default font size used was incorrect. It is not just 10px, 
but should be based off a percentage of the video size.
b) The top/left CSS property needs to be adjusted appropriately if the cue
is center aligned and we change the cue's size such that the cue remains centered.
* html/track/TextTrackCueGeneric.cpp:
(WebCore::TextTrackCueGenericBoxElement::applyCSSProperties):
* html/track/VTTCue.cpp:
(WebCore::VTTCueBox::applyCSSProperties):
* html/track/VTTCue.h: Remove unnecessary constant.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@172421 268f45cc-cd09-0410-ab3c-d52691b4dbfc