[EFL] Bump up and patch dependencies to allow building for aarch64 https://bugs.webkit.org/show_bug.cgi?id=135885
Patch by Akos Kiss <akiss@inf.u-szeged.hu> on 2014-08-13
Reviewed by Gyuyoung Kim.

The currently used versions of pixman, libffi, glib, glib-networking,
and fontconfig fail to configure for aarch64, so they are bumped up to
the latest stable version.

Gstreamer-plugins-base configures for aarch64 but incorrectly recogizes
ARM NEON assembly support, so it is patched with a change that has been
accepted by upstream.

* efl/jhbuild.modules:
* efl/patches/gst-prevent-neon-check-in-configure-from-passing-under-aarch64.patch: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@172549 268f45cc-cd09-0410-ab3c-d52691b4dbfc