run-webkit-tests should skip 'support' directories https://bugs.webkit.org/show_bug.cgi?id=135910
Patch by Rebecca Hauck <rhauck@adobe.com> on 2014-08-15
Reviewed by Bem Jones-Bey.

This patch adds the 'support' directory to the list
of dirs to skip when looking for runnable tests.
the W3C uses this name for support files, so tests
imported from the W3C will not work without this change
or without many edits.

* Scripts/webkitpy/port/base.py:
(Port._real_tests):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@172638 268f45cc-cd09-0410-ab3c-d52691b4dbfc