REGRESSION (r164173): Safari AutoFill button in popover, default buttons on iCloud control panel and Safari Safe Browsing pages don't have default appearance https://bugs.webkit.org/show_bug.cgi?id=136011
Reviewed by Sam Weinig.

When we render some controls, we need to make sure the NSWindow subclass we
use as the host of the AppKit views identifies itself as the key window (if
the control state is correct). We already have a flag tracking this for
key appearance - just make sure we do the same for isKeyWindow.

* platform/mac/ThemeMac.mm:
(-[WebCoreThemeWindow isKeyWindow]): New override to mirror the static
variable for key appearance.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@172693 268f45cc-cd09-0410-ab3c-d52691b4dbfc